Moved prompt generation to its own endpoint
